feat: 페이지네이션 getCursor 메서드 processor로 분리
